feat(python): produce code coverage from a unit test
