+++ MODULE.bazel
@@ -9,7 +9,7 @@ bazel_dep(name = "apple_support", version = "1.15.1")
 bazel_dep(name = "aspect_bazel_lib", version = "2.10.0")
 bazel_dep(name = "aspect_rules_js", version = "2.0.0")
 bazel_dep(name = "aspect_rules_lint", version = "1.0.0-rc10")
-bazel_dep(name = "aspect_rules_py", version = "1.0.0")
+bazel_dep(name = "aspect_rules_py", version = "1.2.1")
 bazel_dep(name = "aspect_rules_swc", version = "2.0.0")
 bazel_dep(name = "aspect_rules_ts", version = "3.0.0")
 bazel_dep(name = "bazel_skylib", version = "1.7.1")
